Time: 59 ms (88.00%) | Memory: 47.9 MB (69.33%) - LeetSync

+++ 3974-xor-after-range-multiplication-queries-i/xor-after-range-multiplication-queries-i.java
@@ -2,7 +2,7 @@ class Solution {
 
 
     public int xorAfterQueries(int[] nums, int[][] queries) {
-     int mod = 1000000007;
+     long mod = 1000000007;
         // Process each query
         for (int[] t : queries) {
             int l = t[0];
